+++ TShockAPI/TSPlayer.cs
@@ -1394,7 +1394,9 @@ namespace TShockAPI
 		/// <param name="tiley">The Y coordinate.</param>
 		/// <param name="context">The PlayerSpawnContext.</param>
 		/// <param name="respawnTimer">The respawn timer, will be Player.respawnTimer if parameter is null.</param>
-		public void Spawn(int tilex, int tiley, PlayerSpawnContext context, int? respawnTimer = null)
+		/// <param name="numberOfDeathsPVE">The number of deaths PVE, will be TPlayer.numberOfDeathsPVE if parameter is null.</param>
+		/// <param name="numberOfDeathsPVP">The number of deaths PVP, will be TPlayer.numberOfDeathsPVP if parameter is null.</param>
+		public void Spawn(int tilex, int tiley, PlayerSpawnContext context, int? respawnTimer = null, short? numberOfDeathsPVE = null, short? numberOfDeathsPVP = null)
 		{
 			using (var ms = new MemoryStream())
 			{
@@ -1404,6 +1406,10 @@ namespace TShockAPI
 					TileX = (short)tilex,
 					TileY = (short)tiley,
 					RespawnTimer = respawnTimer ?? TShock.Players[Index].RespawnTimer * 60,
+
+					numberOfDeathsPVE = numberOfDeathsPVE ?? (short)TPlayer.numberOfDeathsPVE,
+					numberOfDeathsPVP = numberOfDeathsPVP ?? (short)TPlayer.numberOfDeathsPVP,
+
 					PlayerSpawnContext = context,
 				};
 				msg.PackFull(ms);